Squashed commit of the following:
commit 42274ec82a5dcbd1b6b0bc3052bbdf3f83960315
Author: babisRoutis <haralampos.routis@netcompany.com>
Date:   Sat Jun 20 23:55:13 2026 +0300

    Removed unused serialization classes and enums, refactored `IssueCredential` and `GetDeferredCredential` to simplify error handling, and centralized related type definitions.

commit 266b6853d5de99aadfea5d6c3b9945c290971acc
Author: babisRoutis <haralampos.routis@netcompany.com>
Date:   Sat Jun 20 23:10:14 2026 +0300

    Added `RequestEncryptionError` and `IssueCredentialResponse` implementations and refactored `IssueCredential` to improve encryption error handling and streamline request validation logic.

commit f7f71277ddef71ca9b46a50310b8a2511e0af262
Author: babisRoutis <haralampos.routis@netcompany.com>
Date:   Sat Jun 20 19:12:48 2026 +0300

    Removed deprecated and unused code, including `@Deprecated` annotation in `CredentialIssuerMetaData`, `Either` reference in `CredentialRequest`, and unused imports. Updated `AuthorizedCredentialRequest` to replace `id` with `credentialId`. Added `@Suppress("unused")` to unused enum in `CreateCredentialsOffer`. Streamlined request handling in `IssueCredential`.

commit cd62ea1a94e1ec023c1858dc93595d7f411fbf2c
Author: babisRoutis <haralampos.routis@netcompany.com>
Date:   Sat Jun 20 19:08:23 2026 +0300

    Replaced `requestByCredentialIdentifier` with `requestByCredentialConfigurationId` across `WalletApiTest` and related tests. Updated test logic to reflect changes in credential request handling. Simplified `authorizeRequest` in `IssueCredential` by removing suspend modifier. Removed unused functions and test methods for `requestByCredentialIdentifier`.

commit c92e76a8ed49ac6e5d1e977c0c071cb3a10aae81
Author: babisRoutis <haralampos.routis@netcompany.com>
Date:   Sat Jun 20 18:45:12 2026 +0300

    Removed unused `DefaultResolveCredentialRequestByCredentialIdentifier` and related classes, functions, and imports. Refactored `IssueCredential` logic to remove dependency on deprecated resolution mechanism. Simplified authorization and request handling workflows.
